+++ demisto_sdk/commands/generate_docs/tests/generate_docs_test.py
@@ -503,3 +503,8 @@ class TestGenerateIntegrationDoc:
         generate_integration_doc(TEST_INTEGRATION_PATH)
         assert open(fake_readme).read() == open(
             os.path.join(os.path.dirname(TEST_INTEGRATION_PATH), 'README.md')).read()
+
+        # Test that the predefined values and default values are added to the README.
+        assert "The type of the newly created user The optional values are ['Basic', 'Pro', 'Corporate']. " \
+               "The default value is Basic." in open(fake_readme).read()
+        assert "Number of users to return. Max 300. The default value is 30." in open(fake_readme).read()